New Config; Better Handling
Added new configuration system.
Rebuilt command and event handling systems.
Cleaned up.
Changed some scripts.
