+++ cache/kick.ts
@@ -1,23 +1,21 @@
-import infraction from '@/database/schemas/infraction';
+/*import infraction from '@/database/schemas/infraction';
 import { CommandInfo, CommandRun } from '@/Interfaces';
 import gen from '@/utils/gen';
 import getUser from '@/utils/getUser';
 import { MessageEmbed, TextChannel } from 'discord.js';
-import ms from 'ms';
 
 export const run: CommandRun = async (client, message, args) => {
 	const member = await getUser(message, message.mentions.users.first() || args[0]);
 	if (!(await member.moderatable)) return message.channel.send({ content: 'I can not moderate this user.' });
-	const reason = message.content.split(/ +/g).slice(3).join(' ');
+	const reason = message.content.split(/ +/g).slice(2).join(' ');
 	const moderator = message.member;
-	const time = args[1];
 	if (!reason) return message.reply({ content: 'You did not specify a reason!' });
 	const infID = await gen('id', 6);
 	let userEmbed = new MessageEmbed({
-		title: 'You have been timed-out in FyreNodes.',
+		title: 'You have been kicked from FyreNodes.',
 		thumbnail: { url: message.guild.iconURL() },
-		color: '#FA8A29',
-		description: `**ID:** ${infID}\n**Reason:** ${reason.replace('-s', '')}\n**Duration:** ${ms(ms(time), { long: true })}\n${message.content.includes('-s') ? '' : `**Moderator:** ${moderator.user.tag} (<@${moderator.id}>)`}`,
+		color: '#FA2929',
+		description: `**ID:** ${infID}\n**Reason:** ${reason.replace('-s', '')}\n${message.content.includes('-s') ? '' : `**Moderator:** ${moderator.user.tag} (<@${moderator.id}>)`}`,
 		footer: {
 			text: 'Jarvis Moderation',
 			iconURL: client.user.avatarURL()
@@ -26,10 +24,10 @@ export const run: CommandRun = async (client, message, args) => {
 	});
 	let logsEmbed = new MessageEmbed({
 		author: {
-			name: `Infraction | Timeout | ${infID}`,
+			name: `Infraction | Kick | ${infID}`,
 			iconURL: member.user.avatarURL()
 		},
-		color: '#FA8A29',
+		color: '#FA2929',
 		fields: [
 			{
 				name: 'User:',
@@ -41,11 +39,6 @@ export const run: CommandRun = async (client, message, args) => {
 				value: `${moderator.user.tag} (<@${moderator.id}>)`,
 				inline: true
 			},
-			{
-				name: 'Duration:',
-				value: ms(ms(time), { long: true }),
-				inline: true
-			},
 			{ name: 'Reason:', value: reason, inline: false }
 		],
 		footer: {
@@ -54,25 +47,21 @@ export const run: CommandRun = async (client, message, args) => {
 		},
 		timestamp: Date.now()
 	});
-	await member.timeout(ms(time), `Moderator: ${moderator.user.tag}; Reason: ${reason}`);
 	await infraction.create({
 		infID: infID,
 		guild: message.guild.id,
-		type: 'Timeout',
+		type: 'Kick',
 		user: member.id,
-		details: {
-			reason: reason,
-			moderator: moderator.id,
-			duration: ms(ms(time), { long: true })
-		}
+		details: { reason: reason, moderator: moderator.id }
 	});
 	await member.send({ embeds: [userEmbed] }).catch(() => {});
+	await member.kick(`Moderator: ${moderator.user.tag}; Reason: ${reason}`);
 	await (message.guild.channels.cache.get('968665856919888002') as TextChannel).send({ embeds: [logsEmbed] });
-	message.channel.send({ content: `Successfully timed-out ${member.user.tag} for ${ms(ms(time), { long: true })}.` });
+	message.channel.send({ content: `Successfully kicked ${member.user.tag}.` });
 };
 
 export const info: CommandInfo = {
-	name: 'timeout',
+	name: 'kick',
 	category: 'moderation',
-	permissions: ['MODERATE_MEMBERS']
-};
+	permissions: ['KICK_MEMBERS']
+};*/
