+++ cache/timeout.ts
@@ -1,21 +1,23 @@
-import infraction from '@/database/schemas/infraction';
+/*import infraction from '@/database/schemas/infraction';
 import { CommandInfo, CommandRun } from '@/Interfaces';
 import gen from '@/utils/gen';
 import getUser from '@/utils/getUser';
 import { MessageEmbed, TextChannel } from 'discord.js';
+import ms from 'ms';
 
 export const run: CommandRun = async (client, message, args) => {
 	const member = await getUser(message, message.mentions.users.first() || args[0]);
 	if (!(await member.moderatable)) return message.channel.send({ content: 'I can not moderate this user.' });
-	const reason = message.content.split(/ +/g).slice(2).join(' ');
+	const reason = message.content.split(/ +/g).slice(3).join(' ');
 	const moderator = message.member;
+	const time = args[1];
 	if (!reason) return message.reply({ content: 'You did not specify a reason!' });
 	const infID = await gen('id', 6);
 	let userEmbed = new MessageEmbed({
-		title: 'You have been banned from FyreNodes.',
+		title: 'You have been timed-out in FyreNodes.',
 		thumbnail: { url: message.guild.iconURL() },
-		color: '#FA2929',
-		description: `**ID:** ${infID}\n**Reason:** ${reason.replace('-s', '')}\n${message.content.includes('-s') ? '' : `**Moderator:** ${moderator.user.tag} (<@${moderator.id}>)`}`,
+		color: '#FA8A29',
+		description: `**ID:** ${infID}\n**Reason:** ${reason.replace('-s', '')}\n**Duration:** ${ms(ms(time), { long: true })}\n${message.content.includes('-s') ? '' : `**Moderator:** ${moderator.user.tag} (<@${moderator.id}>)`}`,
 		footer: {
 			text: 'Jarvis Moderation',
 			iconURL: client.user.avatarURL()
@@ -24,10 +26,10 @@ export const run: CommandRun = async (client, message, args) => {
 	});
 	let logsEmbed = new MessageEmbed({
 		author: {
-			name: `Infraction | Ban | ${infID}`,
+			name: `Infraction | Timeout | ${infID}`,
 			iconURL: member.user.avatarURL()
 		},
-		color: '#FA2929',
+		color: '#FA8A29',
 		fields: [
 			{
 				name: 'User:',
@@ -39,6 +41,11 @@ export const run: CommandRun = async (client, message, args) => {
 				value: `${moderator.user.tag} (<@${moderator.id}>)`,
 				inline: true
 			},
+			{
+				name: 'Duration:',
+				value: ms(ms(time), { long: true }),
+				inline: true
+			},
 			{ name: 'Reason:', value: reason, inline: false }
 		],
 		footer: {
@@ -47,21 +54,25 @@ export const run: CommandRun = async (client, message, args) => {
 		},
 		timestamp: Date.now()
 	});
+	await member.timeout(ms(time), `Moderator: ${moderator.user.tag}; Reason: ${reason}`);
 	await infraction.create({
 		infID: infID,
 		guild: message.guild.id,
-		type: 'Ban',
+		type: 'Timeout',
 		user: member.id,
-		details: { reason: reason, moderator: moderator.id }
+		details: {
+			reason: reason,
+			moderator: moderator.id,
+			duration: ms(ms(time), { long: true })
+		}
 	});
 	await member.send({ embeds: [userEmbed] }).catch(() => {});
-	await member.ban({ days: 3, reason: `Moderator: ${moderator.user.tag}; Reason: ${reason}` });
 	await (message.guild.channels.cache.get('968665856919888002') as TextChannel).send({ embeds: [logsEmbed] });
-	message.channel.send({ content: `Successfully banned ${member.user.tag}.` });
+	message.channel.send({ content: `Successfully timed-out ${member.user.tag} for ${ms(ms(time), { long: true })}.` });
 };
 
 export const info: CommandInfo = {
-	name: 'ban',
+	name: 'timeout',
 	category: 'moderation',
-	permissions: ['BAN_MEMBERS']
-};
+	permissions: ['MODERATE_MEMBERS']
+};*/
