+++ package.json
@@ -2,14 +2,14 @@
 	"name": "jarvis",
 	"packageManager": "yarn@3.2.0",
 	"description": "Multipurpose bot for FyreNodes Discord server.",
-	"version": "0.2.0",
+	"version": "0.3.0",
 	"main": "index.ts",
 	"engines": {
 		"node": ">=16.6.0"
 	},
 	"scripts": {
 		"start": "tsc --build; tsc-alias -p tsconfig.json; NODE_ENV=production node -r dotenv/config dist/index.js",
-		"dev": "clear; NODE_ENV=development ts-node-dev -r tsconfig-paths/register -r dotenv/config index.ts",
+		"dev": "clear; NODE_ENV=development ts-node-dev -r tsconfig-paths/register -r dotenv/config --project tsconfig.json index.ts",
 		"lint": "prettier --check 'src/**/*'",
 		"lint:fix": "prettier --write src/**/*"
 	},
@@ -41,7 +41,8 @@
 		"@types/ms": "^0.7.31",
 		"@types/node": "^17.0.26",
 		"prettier": "2.6.2",
-		"ts-node-dev": "^1.1.8",
+		"ts-node": "^10.8.0",
+		"ts-node-dev": "^2.0.0",
 		"tsconfig-paths": "^3.14.1",
 		"typescript": "^4.6.3"
 	}
