+++ src/Client.ts
@@ -1,29 +1,21 @@
-import config, { JarvisConfig } from '@config';
-import { Client as DiscordClient, Collection, User } from 'discord.js';
-import { Command, Interaction } from '@/Interfaces';
-import { commandHandler, interactionHandler, eventHandler } from '@/Handlers';
-import connect from '@/database/connect';
-import transcripts from '@/helpers/transcripts';
-import database from './helpers/database';
-import dayjs from './helpers/modules/dayjs';
+import config from '@config';
+import { Client as DiscordClient, Collection, MessageOptions, TextChannel, User } from 'discord.js';
+import { Command, BaseCommand, JarvisConfig, Button } from '@/Interfaces';
+import { LogChannel, PermissionLevel } from '@/interfaces/Config';
 
 export default class Client extends DiscordClient {
 	public commands: Collection<string, Command> = new Collection();
-	public interactions: Collection<string, Interaction> = new Collection();
+	public buttons: Collection<string, Button> = new Collection();
+	public baseCommands: Collection<string, BaseCommand> = new Collection();
 	public config: JarvisConfig = config;
 
-	public async init() {
-		await transcripts();
-		eventHandler(this);
-		commandHandler(this);
-		interactionHandler(this);
-		await database();
-		dayjs();
-		connect();
-		this.login(process.env.TOKEN);
+	public getPermissionLevel = (user: User): PermissionLevel => {
+		if (!this.config.perms.includes({ id: user.id })) return 0;
+		return this.config.perms.find((x) => x.id === user.id).level;
 	};
 
-	public isDeveloper(user: User): boolean {
-		if (config.developers.includes(user.id)) return true; else return false;
+	public log = (ch: LogChannel, msg: MessageOptions) => {
+		const channel = this.channels.cache.find(Object.keys(this.config.channels.logs)[ch]) as TextChannel;
+		channel.send(msg);
 	};
-}
+};
