+++ src/Handlers.ts
@@ -1,3 +1,3 @@
+export { baseCommandHandler } from '@/handlers/baseCommand';
 export { commandHandler } from '@/handlers/command';
-export { interactionHandler } from '@/handlers/interaction';
 export { eventHandler } from '@/handlers/event';
