+++ src/Interfaces.ts
@@ -1,2 +1,4 @@
-export { Command, CommandRun, CommandInfo } from '@/interfaces/Command';
-export { Interaction, InteractionRun, InteractionInfo } from '@/interfaces/Interaction';
+export { Command, CommandInfo, CommandRun } from '@/interfaces/Command';
+export { BaseCommand, BaseCommandInfo, BaseCommandRun } from '@/interfaces/BaseCommand';
+export { JarvisConfig } from '@/interfaces/Config';
+export { Button } from '@/interfaces/Button';
