+++ src/base/utils/test.ts
@@ -1,0 +1,11 @@
+import { BaseCommandInfo, BaseCommandRun } from '@/Interfaces';
+import { MessageActionRow, Modal, TextInputComponent } from 'discord.js';
+
+export const run: BaseCommandRun = async (client, message, args) => {
+	
+};
+
+export const info: BaseCommandInfo = {
+	name: 'test',
+	category: 'test'
+};
