+++ src/commands/moderation/clear.ts
@@ -1,21 +1,31 @@
-import { CommandInfo, CommandRun } from '@/Interfaces';
-import { TextChannel } from 'discord.js';
-import ms from 'ms';
+import { CommandInfo, CommandRun } from "@/Interfaces";
+import { CommandInteraction } from "discord.js";
+import ms from "ms";
 
-export const run: CommandRun = async (client, message, args) => {
-	if (isNaN(parseInt(args[0])) || parseInt(args[0]) > 100 || parseInt(args[0]) < 1) return message.reply({ content: 'Please specify an amout between 1-100.' });
-	const messages = await message.channel.messages.fetch({ limit: parseInt(args[0]) });
-	const usable = messages.filter((m) => m.createdTimestamp - Date.now() < ms('14d') && !m.pinned);
-	await (message.channel as TextChannel).bulkDelete(usable);
-	message.channel.send({ content: 'Successfully deleted messages.' }).then(async (msg) => {
-		setTimeout(async () => {
-			await msg.delete();
-		}, 3000);
-	});
+export const run: CommandRun = async (client, interaction: CommandInteraction) => {
+    const messages = await interaction.channel.messages.fetch({ limit: interaction.options.getNumber('amount') });
+    const useable = messages.filter((m) => m.createdTimestamp - Date.now() < ms('14d') && !m.pinned);
+    await interaction.channel.bulkDelete(useable);
+    await interaction.reply({ content: 'Successfully deleted messages.' });
+    setTimeout(async () => {
+        await interaction.deleteReply();
+    }, 3000);
 };
 
 export const info: CommandInfo = {
-	name: 'clear',
-	category: 'moderation',
-	permissions: ['MANAGE_CHANNELS']
+    name: 'clear',
+    category: 'moderation',
+    description: 'Purges messages in current channel',
+    default_member_permissions: 0x10,
+    dm_permission: false,
+    options: [
+        {
+            type: 10,
+            name: 'amount',
+            description: 'Amount of messages to purge',
+            required: true,
+            min_value: 1,
+            max_value: 100
+        }
+    ]
 };
