+++ src/commands/tickets/close.ts
@@ -1,15 +1,14 @@
-import { InteractionInfo, InteractionRun } from '@/Interfaces';
+import { CommandInfo, CommandRun } from '@/Interfaces';
 import { CommandInteraction } from 'discord.js';
 import closeTicket from '@/functions/closeTicket';
 
-export const run: InteractionRun = async (client, interaction: CommandInteraction) => {
+export const run: CommandRun = async (client, interaction: CommandInteraction) => {
 	await closeTicket(client, interaction);
 };
 
-export const info: InteractionInfo = {
+export const info: CommandInfo = {
 	name: 'close',
 	category: 'tickets',
 	description: 'Close a support ticket.',
-	intType: 'command',
-	type: 1
+	dm_permission: false
 };
