+++ src/commands/tickets/open.ts
@@ -1,25 +1,25 @@
 import ticket from '@/database/schemas/ticket';
-import { InteractionInfo, InteractionRun } from '@/Interfaces';
+import { CommandInfo, CommandRun } from '@/Interfaces';
 import gen from '@/utils/gen';
 import { CommandInteraction, MessageActionRow, MessageButton, MessageEmbed, Permissions } from 'discord.js';
 
-export const run: InteractionRun = async (client, interaction: CommandInteraction) => {
+export const run: CommandRun = async (client, interaction: CommandInteraction) => {
 	if (await ticket.exists({ guild: interaction.guild.id, user: interaction.user.id, status: 'open' })) return interaction.reply({ content: 'You already have an open ticket.', ephemeral: true });
 	const id = await gen('id', 6);
 	const ch = await interaction.guild.channels.create(`${interaction.options.get('department').value}-${interaction.user.username}`, {
 		type: 'GUILD_TEXT',
-		parent: '974147300094009384',
+		parent: client.config.tickets,
 		topic: `${interaction.options.get('department').value.toString().replace('billing', 'Billing').replace('tech', 'Technical').replace('general', 'General')} Support ticket for ${interaction.user.tag} (${interaction.user.id})`,
 		reason: 'Automated Action: Ticket Created',
 		position: 1,
 		permissionOverwrites: [
 			{
-				id: '961173678249357336',
+				id: client.config.roles.support,
 				type: 'role',
 				allow: [Permissions.FLAGS.VIEW_CHANNEL, Permissions.FLAGS.SEND_MESSAGES, Permissions.FLAGS.ATTACH_FILES, Permissions.FLAGS.EMBED_LINKS, Permissions.FLAGS.ADD_REACTIONS, Permissions.FLAGS.READ_MESSAGE_HISTORY, Permissions.FLAGS.USE_EXTERNAL_EMOJIS, Permissions.FLAGS.USE_EXTERNAL_STICKERS]
 			},
 			{
-				id: '961172814063337523',
+				id: client.config.roles.admin,
 				type: 'role',
 				allow: [
 					Permissions.FLAGS.VIEW_CHANNEL,
@@ -65,12 +65,11 @@ export const run: InteractionRun = async (client, interaction: CommandInteractio
 	ch.send({ embeds: [embed], components: [actionRow] }).then((msg) => msg.pin());
 };
 
-export const info: InteractionInfo = {
+export const info: CommandInfo = {
 	name: 'open',
 	category: 'tickets',
 	description: 'Opens a support ticket.',
-	intType: 'command',
-	type: 1,
+	dm_permission: false,
 	options: [
 		{
 			type: 3,
