--- src/commands/utils/test.ts
@@ -1,11 +1,0 @@
-import { CommandInfo, CommandRun } from '@/Interfaces';
-import { MessageActionRow, Modal, TextInputComponent } from 'discord.js';
-
-export const run: CommandRun = async (client, message, args) => {
-	
-};
-
-export const info: CommandInfo = {
-	name: 'test',
-	category: 'test'
-};
