+++ src/events/interactions/buttonInteraction.ts
@@ -1,10 +1,11 @@
 import Client from '@/Client';
-import { Interaction } from '@/Interfaces';
+import { Button } from '@/Interfaces';
 import { ButtonInteraction } from 'discord.js';
 
 export default async (client: Client, interaction: ButtonInteraction) => {
+	if (!interaction.inGuild) return;
 	const int: string = interaction.customId.toString().toLowerCase();
-	const buttonInteraction: Interaction = await client.interactions.get(int);
+	const buttonInteraction: Button = await client.buttons.get(int);
 	if (!buttonInteraction) return;
 	buttonInteraction.run(client, interaction);
 };
