+++ src/handlers/command.ts
@@ -6,10 +6,10 @@ import { grey, green, white } from 'chalk';
 export const commandHandler = async (client: Client) => {
 	const path: string = `${__dirname}/../commands`;
 	await readdirSync(path).forEach(async (dir: string) => {
-		const commands: string[] = readdirSync(`${path}/${dir}`).filter((file: string) => file.endsWith('.ts') || file.endsWith('js'));
-		for (const cmd of commands) {
-			const command: Command = require(`${path}/${dir}/${cmd}`);
-			await client.commands.set(command.info.name, command);
+		const commands: string[] = await readdirSync(`${path}/${dir}`).filter((file: string) => file.endsWith('.ts') || file.endsWith('.js'));
+		for (let command of commands) {
+			const cmd: Command = require(`${path}/${dir}/${command}`);
+			await client.commands.set(cmd.info.name, cmd);
 		}
 	});
 	console.log(`${grey.bold('[')}${green.bold('HANDLER')}${grey.bold(']')} ${white('All commands have been loaded.')}`);
