+++ src/handlers/event.ts
@@ -10,6 +10,6 @@ export const eventHandler = (client: Client) => {
 	client.once('ready', () => ready(client));
 	client.on('messageCreate', async (message: Message) => messageCreate(client, message));
 	client.on('interactionCreate', async (interaction: Interaction) => interactionCreate(client, interaction));
-	client.on('guildMemberAdd', async (member: GuildMember) => guildMemberAdd(member));
+	client.on('guildMemberAdd', async (member: GuildMember) => guildMemberAdd(client, member));
 	console.log(`${grey.bold('[')}${green.bold('HANDLER')}${grey.bold(']')} ${white('All events have been loaded.')}`);
 };
