+++ src/interfaces/Command.ts
@@ -1,5 +1,5 @@
 import Client from '@/Client';
-import { Message, PermissionFlags, PermissionResolvable, Permissions } from 'discord.js';
+import { CommandInteraction } from 'discord.js';
 
 export interface Command {
 	info: CommandInfo;
@@ -7,11 +7,30 @@ export interface Command {
 }
 
 export interface CommandRun {
-	(client: Client, message: Message, args: string[]);
+	(client: Client, interaction: CommandInteraction);
 }
 
 export interface CommandInfo {
 	name: string;
+	description: string;
 	category: string;
-	permissions?: PermissionResolvable[];
+	options?: Options[];
+	dm_permission: boolean;
+	default_member_permissions?: number;
+}
+
+interface Options {
+	type?: number;
+	name: string;
+	description: string;
+	required?: boolean;
+	min_value?: number;
+	max_value?: number;
+	autocomplete?: boolean;
+	choices?: Choices[];
+}
+
+interface Choices {
+	name: string;
+	value: string;
 }
